requirements.txt
pandas>=1.3.0
pytest>=6.2.0
pdfplumber>=0.5.0
groq>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
reportlab>=4.0.0
